(MAINT) Add group formatting
Prior to this commit, no formatting existed for groups, meaning that
the only way to display them was to blurt the values of the struct
via %v string interpolation.

This commit:

- Adds the String() function for the Melee, Missile, Move, and
  FightingStrength structs, making their non-data representation
  simpler where needed
- Adds the MarkdownTableEntry() and JSON() functions to the Group
  struct, enabling output of a group in either human or machine
  readable formats
- Updates the play command to return the JSON representation of the
  created group (if the user passes json to the format flag) or a
  markdown table representation of the created group (by default or
  if the user passes human to the format flag).

This improves the readability of groups in the terminal and for any
future external consumers of the data model.