(MAINT) Add vscode debugging config
This commit adds configuration to attach to the project with delve for
interactive debug. This is necessary due to the inclusion of the prompt
TUI interfaces, which take over stdin/stdout.

You can run the delve debugger in headless mode and then tell VSCode to
attach to the process by first running this command:

```bash
dlv debug --headless --api-version=2 --listen=127.0.0.1:8181 . -- play
```

Once delve reports that it is listening, starting the "attach" debug
process will set things moving. When the process ends, so does the delve
session.