(MAINT) Begin using bubbletea
Prior to this commit, the app leveraged the glamour library to
render markdown in the terminal and the promptui library to
handle user prompts.

This commit replaces promptui with the promptkit bubble, which
relies on bubbletea, a TUI framework library maintained by
charm (the same team that maintains glamour). This simplified
the implementation of user prompts *and* improved the UX for
them. Further, as promptkit is based on bubbletea, future
work to unify the theme will "just work" for both.

This commit also ensures that the seed is randomized prior to
rolling for captain traits to ensure they do not repeat endlessly.