(MAINT) Introduce tympan
This commit begins the process of pulling out the reusable components
of the application into a separate package, code-named tympan after
the layer of packing placed between the platen and paper to equalize
the pressure during the printing process. In other words, a helper for
platen, a companion project.

Tympan will eventually form the toolkit on which this application and
others can be built. For now, it merely helps clarify what is generic
and what is specific to Flagrant Factions.

Regarding generics, tympan takes advantage of generics as introduced
in go 1.18 - while still in beta, the project has been tested locally.

Further work will need to be done to get tests into shape.