(MAINT) Update modules for tympan
This commit standardizes the conventions for data files in a module
to use the file *name* as the type of data and always place arrays
of like-data under the entries key. This enabled us to take
advantage of generics to read and parse the arbitrary data, obviating
the requirement for read/parse/cache actions for each data type.