(MAINT) Add the telemetry package to Tympan
This commit places the telemetry package in the Tympan library with
documentation. This package can be used to add open telemetry reporting
to a Tympan application. It currently only supports sending the data to
Honeycomb. It includes a noop build to support notel binary builds which
do not send any telemetry.