(MAINT) Add the state package to Tympan
This commit adds the `state` package to the Tympan library and two child
packages, `instance` and `persona`.

The `state` package adds generic handlers for reading/write yaml state
files from/to disk and defines Stateful and Initializable interfaces,
which are used in implementations of state files.

The `instance` package defines a generic Instance type, which can be
used to store arbitrary data on disk (the data type must be specified by
the application using this feature). It includes helpers and handlers.

The `persona` package defines a generic Persona type, which can be used
to store arbitrary data and settings on disk (both types must be
specified by the application using this feature). It includes helpers
and handlers.