(Maint) Add the `forme` package to Tympan
This commit adds the `forme` package and its three child packages,
`root`, `version`, and `explain` to the Tympan library. These packages
provide ready-to-use cobra commands for Tympan applications, if they
want to use them.

The `root` package returns a root command for cobra which handles
initializing logging and configuration and sets a few persistent flags.

The `version` package returns a cobra command to display the
application's version information using Tympan's metadata struct.

The `explain` package leverages the `dossier` package to return a cobra
command which can be used to display non-reference documentation in the
terminal or online.