(MAINT) Update `tympan` base package
This commit updates the base package of the Tympan library to use the
newly added packages and restructure itself so the base package only
provides exactly what is necessary for creating a Tympan application;
chiefly, the Tympan struct itself, its interface, and the logic/structs
for application configuration.

The cache file and functionality has been removed as it is replaced by
the `module` child package.