(MAINT) Add the embedded module filesystem package
Prior to this commit, the core module data lived in the `modules` folder
at the root of the project.

This commit moves the modules folder into a new `emfs` folder at the
root of the project and creates an `emfs` package there which has a
single function, `GetEmbeddedModulesFS()`, which returns an `embed.FS`
which has the modules folder inside it.

This will allow flfa to take advantage of the newly available `modules`
package from Tympan, ensuring that the core data and scripts needed to
play the game are embedded in the binary itself.

This does not prevent the user from downloading and using a later
version of the core module, it merely makes it available on first run
and without requiring the installer to package the files and handle
placing them appropriately.