(MAINT) Reorganize app layout and update main
Prior to this commit, the project included a `main.go` file in the root
folder. This commit moves it to `cmd/flfa` and updates it to use the
changes outlined in the previous commits.

In particular, this commit removes the root, explain, and version
packages from the `cmd` folder and `main` now uses the formes supplied
by the Tympan library to instantiate those commands for the application.

The `play` command is moved to `cmd/flfa` and updated as needed. It now
walks the player through first time setup instead of creating and
displaying a single group.

This commit also updates the goreleaser build configuration to point to
the new location for build and updates the variables as needed.

The modules in `go.mod` and `go.sum` have been updated to reflect the
changes in dependencies and updated to the latest compatible versions.

This commit also moves the build script into the more idiomatic
`scripts` folder in the project root and no longer copies the data or
documentation into the `dist` folder on build.

This commit adds a new script, `debug.ps1`, which can be used to start
a headless delve debugging session so that a developer can debug the
application even when using bubbletea (as prompts break stdio while they
run).