(PROTOTYPE) Update traits & implement scripting
Prior to this commit, the traits in the core module had prototype pokes
at how to define and call traits, but the scripting itself was not
implemented.

This commit adds minimal scripting to cover the addition and removal of
captain and special traits for Groups. In particular, the `core` module
now provides:

- the `dice` standalone library for rolling/rerolling dice
- the core module library, which itself currently only exports one
  submodule, `Group` (called in a tengo script as `core.Group`)
- the group submodule which is a convenience layer for calling the
  `Profile` and `Addenda` submodules
- the `Profile` submodule includes numerous functions for checking and
  modifying a Group, particularly their pre-play definition.
- the `Addenda` submodule, which exports functions for interacting with
  the addenda field of a group - a grab-bag map for holding dynamic
  values that can't be predicted/included in a struct.

Additionally, this commit modifies the trait scripting definitions to
accommodate the actual prototype scripts.