(TYMPAN) Add Printers packages
This commit adds the new printers package to Tympan with two child
packages, `json` and `terminal`. Both of them are intended to provide a
simplified/helpful UX for converting arbitrary data to another format.

Both packages use the options model to allow a user to pass semantically
meaningful sets of options instead of needing to understand all of the
underlying methods.

The `StructToJson` function in particular is useful for being able to
decide at conversion time whether or not the json should be compressed
or if any particular keys should be dropped.

The primary benefit of the terminal package is a convenient way to pass
and instantiate shared styles and settings between commands and to allow
you to pass arbitrary flags to change display behavior.