(TYMPAN) Add prompts module & packages
This commit adds the prompts module to Tympan and includes three
packages in it, all of which provide a friendlier UX over Erik Geiser's
incredible promptkit module:

- the `confirmer` package maps to promptkit's confirmation package
- the `texter` package maps to promptkit's textinput package
- the `selector` package maps to promptkit's selection package

These packages include alternate templates and functions for passing
options when creating a prompt so a developer doesn't need to read the
docs as extensively.

The texter module also provides helpers for creating validatable input
text prompts. The selector module provides helpers for creating
selection prompts where the choices are structs, providing options for
viewing the choices as either a list or a table.