(PROTOTYPE) Add initial TUI models
This commit adds the tui module to FLFA, which defines a basic shared
model and enables any model to save the current configuration.

It also entirely replaces the previous serial prompt model with the new
bubbletea approach, putting forward three new models: persona, group,
and company.

This commit lays the groundwork for being able to create and edit groups
and companies for further use and provides a maintainable structure for
definining TUI models and prompts.