(MAINT) Move engine initialization logic
This commit moves the initialization logic into the InitializeEngine method
on the API instead of in a function in a child package. This is the only
place that function was used.

This reduces the need to look through files and removes the name collision
between the FLFA scripting package and the Tympan scripting package.