(MAINT) Rename Persona to Player
This commit renames the persona package in FLFA to player and creates a
Player type to hold the generic implementation of persona to remove the
need to specify the type parameters everywhere. This also reflects a
more correct semantic for the persona type. These changes do not impact
the names of any methods or functions, only the parameters and logic to
ensure the code works after the change.