(TYMPAN) Handle submodel behavior with end funcs
This commit updates the end functions for compositor models, returning an EndMsg
if it is a submodel and tea.Quit if it is not. It also updates the Broken() end
function to set the state of the model to bring it in line with the behavior of
the other end functions.