(MAINT) Update Player caching/retrieval
This commit updates the caching logic for retrieving players to replace
an existing player in the cache if there is a newer version on disk and
append only if the player is not already cached. It also updates the
GetPlayer function to return a Player object instead of the underlying
persona object.