(GH-8) Make Editor UI look like an app
Prior to this commit, the editor UI butted up against the top left of the terminal
and had no custom styling to indicate that an app was running beyond the prompt
messages. This commit updates the styling so the app displays centered in a bordered
box with the app title and subtitle at the top.

It introduces two new functions and one new method to the TUI library:

- The TerminalSettings function includes standard styles and colors for the app so
  we do not need to recreate the settings everywhere.
- The Title function creates the app title, centered and colored.
- The Display method pulls the view body and title together, placing them inside
  a box with a colored border and centering the view.