(TYMPAN) Fix end methods to actually quit
Prior to this commit, using the Done and Cancel methods to exit a program behaved
strangely, requiring an additional key press to exit. This is because they were
returning the tea.Quit *function* instead of the quit message that function
returns, so the program never actually quit and only exited because of an
unhandled keypress.

This commit updates the end methods of the compositor to actually call tea.Quit()
instead of returning the function, ensuring the program exits.