(GH-8) Handle cancels and replacement in player menu
Prior to this commit, the player model did not handle pressing the escape key
in any of the menus, meaning users could not cancel or back out of them. This
commit implements handling for cancelling menus, moving to the next appropriate
menu. If cancelling a prompt would end the model, the implementation is aware
of whether or not it is run as a submodel; if it is not, cancelling a main or
critical menu returns tea.Quit. If it is a submodel, it instead returns cancel.

This commit also modifies the implementation to handle replacing the player
persona the user is editing in the parent menu gracefully.