(GH-8) Make display more dynamic to buffer size
This commit updates the display method on the shared model to render
the app view with more awareness of the size of the buffer to improve
how the app displays in different terminals and no longer requiring
terminals of a precise size.

A future commit will need to address the sizing for the prompts.