(TYMPAN) Extend Terminal to support dynamic styles
Prior to this commit, developers could use terminal settings to pull together the
necessary information to style their text but it was clunky and there was no good
way to define compositional styles which built on the extra styles and/or colors.

This commit extends the definition of the Settings struct to include a list of
dynamic styles as a map of operations to compile that style. Because the operations
are run when they are called to apply, this allows a developer to compose complex
styles and modify the settings before calling them.

This should enable some refactoring to support the shift towards a TUI theme that
can be extended/overwritten from data to enable module theming.