(TYMPAN) Ensure  copying terminal settings copies all rules and styles
Prior to this commit, the implementation of the Copy method for terminal Settings
failed to copy over the styles at all and colors in the extra map. This was due
to the nature of encoding/decoding private fields on structs (it does not work)
and the infeasibility of replacing a pointer variable in child scope func in a
loop without actually reassigning the variable in the parent scope.

This commit updates the signature of a terminal settings option to also return
the settings object so that it can be reassigned, which solved half of the copy
problem. The other half is addressed by reimplementing copy to use copystructure
and a short loop to clone the styles explicitly, since no other method works.

Finally, this commit adds the new From option, which developers can use when creating
terminal settings to start from the settings of an existing instance. This is most
useful for when you want to provide defaults for a model but also enable folks to
override/extend without having to manually update every style one after the other.