(TYMPAN) Add Terminal Settings to Compositor
This commit adds terminal settings to the Compositor so they can be used for styling.

It adds the DefaultTerminalSettings method to define the terminal settings needed for
styling the view of a fatal error, updates the ViewFatalError method to use the settings,
and adds a New function for creating a Compositor while allowing you to extend/override
the default terminal settings.