(TYMPAN) Improve Flag handling
Prior to this commit, using a flag required retrieving it and comparing its value
to the constants in the terminal package. Additionally, copying terminal settings
broke flags via the copystruct operation.

This commit ensures flags are retained in their correct state during a copy and
adds three new helper methods: FlagIsOn, FlagIsOff, and FlagIsUnset to retrieve
a flag and return true if the flag is on/off/unset respectively, removing the
need to manually compare values every time.