(TYMPAN) Add conditional operations to terminal
Prior to this commit, dynamic styles did not have a way to specify if they should
only apply some operation conditionally, requiring dynamic styles for each possible
case or to add style handling by the caller, defeating the purpose.

This commit adds the concept of conditions to the terminal package, functions which
have access to an instance of terminal Settings and must return true or false. They
are used in the new conditional operations, which all take a condition as input and
only apply the operation if the condition returns true.

This commit adds some default conditions for developers to leverage flags.