(GH-8) Overhaul TUI styling
This commit modifies the TUI styling drastically, taking advantage of the
new dynamic styles and conditional operations from the terminal package to
define settings once and propagate them throughout the application, using
them in various prompts and displays.

This necessitated touching nearly everything with prompts, as the overhaul
required passing the terminal settings to each prompt-builder.