(TYMPAN) Add options to override default color for selector
Prior to this commit, a developer needed to either use the default styling for
selected entries or write their own choice style function.

This commit adds helpers to use the default selection and final choice style
functions but with their own preferred foreground color instead of the default
royal blue.